+++ project/src/components/private-route/private-route.tsx
@@ -2,7 +2,7 @@ import {Route, Redirect} from 'react-router-dom';
 import {RouteProps} from 'react-router-dom';
 import { useSelector} from 'react-redux';
 import {AppRoute, AuthorizationStatus} from '../../const';
-import { getAuthorizationStatus } from '../../store/user-process/selectors';
+import { getAuthorizationStatus } from '../../store/user/selectors';
 
 type PrivateRouteProps = RouteProps & {
   render: () => JSX.Element;
